+++ console/folder/migrate/migration/1475586510_playerspecial.php
@@ -4,7 +4,7 @@ $q = array();
 
 $q[] = 'CREATE TABLE `playerspecial`
         (
-            `playerspecial_level` TINYINT(1) DEFAULT 0,
+            `playerspecial_level` INT(1) DEFAULT 0,
             `playerspecial_player_id` INT(11) DEFAULT 0,
-            `playerspecial_special_id` TINYINT(2) DEFAULT 0
+            `playerspecial_special_id` INT(2) DEFAULT 0
         );';
