+++ console/folder/migrate/migration/1475601521_historytext.php
@@ -4,7 +4,7 @@ $q = array();
 
 $q[] = 'CREATE TABLE `historytext`
         (
-            `historytext_id` TINYINT(2) PRIMARY KEY AUTO_INCREMENT,
+            `historytext_id` INT(2) PRIMARY KEY AUTO_INCREMENT,
             `historytext_name` VARCHAR(255) NOT NULL
         );';
 $q[] = "INSERT INTO `historytext` (`historytext_name`)
