+++ console/folder/migrate/migration/1477287593_news.php
@@ -5,7 +5,7 @@ $q = array();
 $q[] = 'CREATE TABLE `news`
         (
             `news_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
-            `news_country_id` SMALLINT(3) DEFAULT 0,
+            `news_country_id` INT(3) DEFAULT 0,
             `news_date` INT(11) DEFAULT 0,
             `news_text` TEXT NOT NULL,
             `news_title` VARCHAR(255) NOT NULL,
