+++ console/folder/migrate/migration/1477311015_message.php
@@ -6,11 +6,11 @@ $q[] = 'CREATE TABLE `message`
         (
             `message_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
             `message_date` INT(11) DEFAULT 0,
-            `message_delete_from` TINYINT(1) DEFAULT 0,
-            `message_delete_to` TINYINT(1) DEFAULT 0,
-            `message_read` TINYINT(1) DEFAULT 0,
-            `message_support_from` TINYINT(1) DEFAULT 0,
-            `message_support_to` TINYINT(1) DEFAULT 0,
+            `message_delete_from` INT(1) DEFAULT 0,
+            `message_delete_to` INT(1) DEFAULT 0,
+            `message_read` INT(1) DEFAULT 0,
+            `message_support_from` INT(1) DEFAULT 0,
+            `message_support_to` INT(1) DEFAULT 0,
             `message_text` TEXT NOT NULL,
             `message_user_id_from` INT(11) DEFAULT 0,
             `message_user_id_to` INT(11) DEFAULT 0
