+++ console/folder/migrate/migration/1477634387_votestatus.php
@@ -4,7 +4,7 @@ $q = array();
 
 $q[] = 'CREATE TABLE `votestatus`
         (
-            `votestatus_id` TINYINT(1) PRIMARY KEY AUTO_INCREMENT,
+            `votestatus_id` INT(1) PRIMARY KEY AUTO_INCREMENT,
             `votestatus_name` VARCHAR(255) NOT NULL
         );';
 $q[] = "INSERT INTO `votestatus` (`votestatus_name`)
