+++ console/folder/migrate/migration/1479192719_tactic.php
@@ -4,7 +4,7 @@ $q = array();
 
 $q[] = 'CREATE TABLE `tactic`
         (
-            `tactic_id` TINYINT(1) PRIMARY KEY AUTO_INCREMENT,
+            `tactic_id` INT(1) PRIMARY KEY AUTO_INCREMENT,
             `tactic_name` VARCHAR(255) NOT NULL
         );';
 $q[] = "INSERT INTO `tactic` (`tactic_name`)
