+++ console/folder/migrate/migration/1479193019_rude.php
@@ -4,7 +4,7 @@ $q = array();
 
 $q[] = 'CREATE TABLE `rude`
         (
-            `rude_id` TINYINT(1) PRIMARY KEY AUTO_INCREMENT,
+            `rude_id` INT(1) PRIMARY KEY AUTO_INCREMENT,
             `rude_name` VARCHAR(255) NOT NULL
         );';
 $q[] = "INSERT INTO `rude` (`rude_name`)
