+++ console/folder/migrate/migration/1480011107_buildingbase.php
@@ -5,9 +5,9 @@ $q = array();
 $q[] = 'CREATE TABLE `buildingbase`
         (
             `buildingbase_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
-            `buildingbase_building_id` TINYINT(1) DEFAULT 0,
-            `buildingbase_constructiontype_id` TINYINT(1) DEFAULT 0,
-            `buildingbase_day` TINYINT(2) DEFAULT 0,
-            `buildingbase_ready` TINYINT(1) DEFAULT 0,
-            `buildingbase_team_id` SMALLINT(5) DEFAULT 0
+            `buildingbase_building_id` INT(1) DEFAULT 0,
+            `buildingbase_constructiontype_id` INT(1) DEFAULT 0,
+            `buildingbase_day` INT(2) DEFAULT 0,
+            `buildingbase_ready` INT(1) DEFAULT 0,
+            `buildingbase_team_id` INT(5) DEFAULT 0
         );';
