+++ console/folder/migrate/migration/1480445150_school.php
@@ -5,11 +5,11 @@ $q = array();
 $q[] = 'CREATE TABLE `school`
         (
             `school_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
-            `school_day` TINYINT(2) DEFAULT 0,
-            `school_position_id` TINYINT(1) DEFAULT 0,
-            `school_ready` TINYINT(1) DEFAULT 0,
-            `school_season_id` SMALLINT(5) DEFAULT 0,
-            `school_special_id` TINYINT(2) DEFAULT 0,
-            `school_style_id` TINYINT(1) DEFAULT 0,
-            `school_team_id` SMALLINT(5) DEFAULT 0
+            `school_day` INT(2) DEFAULT 0,
+            `school_position_id` INT(1) DEFAULT 0,
+            `school_ready` INT(1) DEFAULT 0,
+            `school_season_id` INT(5) DEFAULT 0,
+            `school_special_id` INT(2) DEFAULT 0,
+            `school_style_id` INT(1) DEFAULT 0,
+            `school_team_id` INT(5) DEFAULT 0
         );';
